requirements.txt
cffi==1.15.1
click==8.1.3
clr-loader==0.1.7
colorama==0.4.5
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
proxy-tools==0.1.0
pycparser==2.21
pythonnet==3.0.0a2
pywebview==3.6.3
Werkzeug==2.2.2
